+++ src/main/java/nl/enjarai/showmeyourskin/client/DummyClientWorld.java
@@ -2,10 +2,12 @@ package nl.enjarai.showmeyourskin.client;
 
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.util.Identifier;
+import net.minecraft.util.registry.BuiltinRegistries;
 import net.minecraft.util.registry.Registry;
 import net.minecraft.util.registry.RegistryEntry;
 import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.Difficulty;
+import net.minecraft.world.dimension.DimensionTypes;
 import nl.enjarai.showmeyourskin.ShowMeYourSkin;
 
 public class DummyClientWorld extends ClientWorld {
@@ -22,7 +24,8 @@ public class DummyClientWorld extends ClientWorld {
                 DummyClientPlayNetworkHandler.getInstance(),
                 new Properties(Difficulty.EASY, false, true),
                 RegistryKey.of(Registry.WORLD_KEY, ShowMeYourSkin.id("dummy")),
-                RegistryEntry.of(DummyDimensionType.getInstance()),
+                BuiltinRegistries.DIMENSION_TYPE.getEntry(DimensionTypes.OVERWORLD)
+                        .orElseThrow(() -> new IllegalStateException("Wait what? Who deleted the overworld?")),
                 0,
                 0,
                 () -> null,