+++ src/main/java/nl/enjarai/showmeyourskin/compat/modmenu/ModMenuIntegration.java
@@ -1,9 +1,10 @@
-package nl.enjarai.showmeyourskin.config;
+package nl.enjarai.showmeyourskin.compat.modmenu;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import nl.enjarai.showmeyourskin.config.ModConfig;
 
 @Environment(EnvType.CLIENT)
 public class ModMenuIntegration implements ModMenuApi {